Move cli_gen/utils.py to base_gen/utils.py
