+++ openapi_spec_tools/base_gen/utils.py
