+++ openapi_spec_tools/cli_gen/layout_generator.py
@@ -1,10 +1,10 @@
 """Declares the LayoutGenerator for inferring a layout from an OpenAPI specification."""
 from typing import Any
 
+from openapi_spec_tools.base_gen.utils import to_snake_case
 from openapi_spec_tools.cli_gen.layout import DEFAULT_START
 from openapi_spec_tools.cli_gen.layout_types import LayoutField
 from openapi_spec_tools.cli_gen.layout_types import LayoutNode
-from openapi_spec_tools.cli_gen.utils import to_snake_case
 from openapi_spec_tools.types import OasField
 
 CREATE = "create"
