+++ tests/base_gen/test_utils.py
@@ -1,13 +1,13 @@
 import pytest
 
-from openapi_spec_tools.cli_gen.utils import is_case_sensitive
-from openapi_spec_tools.cli_gen.utils import maybe_quoted
-from openapi_spec_tools.cli_gen.utils import prepend
-from openapi_spec_tools.cli_gen.utils import replace_special
-from openapi_spec_tools.cli_gen.utils import set_missing
-from openapi_spec_tools.cli_gen.utils import shallow
-from openapi_spec_tools.cli_gen.utils import to_camel_case
-from openapi_spec_tools.cli_gen.utils import to_snake_case
+from openapi_spec_tools.base_gen.utils import is_case_sensitive
+from openapi_spec_tools.base_gen.utils import maybe_quoted
+from openapi_spec_tools.base_gen.utils import prepend
+from openapi_spec_tools.base_gen.utils import replace_special
+from openapi_spec_tools.base_gen.utils import set_missing
+from openapi_spec_tools.base_gen.utils import shallow
+from openapi_spec_tools.base_gen.utils import to_camel_case
+from openapi_spec_tools.base_gen.utils import to_snake_case
 
 
 @pytest.mark.parametrize(
